登录

+++ build.sh
@@ -1,23 +1,6 @@
 #!/bin/bash
 
-function build() {
-    if [ "$3" == "" ];then
-	return;
-    fi
-    cd $1
-    build_dir=$3/$2
-    if [ -d $build_dir ]; then
-	$build_dir/bin/stop.sh
-        rm -rf $build_dir/*
-    fi
-    mvn_tools.sh install
-    tar_file=`find . -name '*.tar.gz'`
-    mkdir -p $build_dir
-    tar zxvf $tar_file -C $build_dir
-    nohup $build_dir/bin/start.sh debug > /dev/null 2>&1 &
-}
-
-function build_view() {
+function build_view_dev() {
     build_dir=/Users/pangxianhai/workspace/recruitment-view
     cd /Users/pangxianhai/projects/clinical-recruitment/recruitment-view
     rm -rf ./dist/*
@@ -27,13 +10,13 @@ function build_view() {
     cd ..
     tar_file=`find . -name '*.tar.gz'`
     if [ -d $build_dir ]; then
-	rm -rf $build_dir
+	  rm -rf $build_dir
     fi
     mkdir -p $build_dir
     tar zxvf $tar_file -C $build_dir    
 }
 
-function build_manager() {
+function build_manager_dev() {
     build_dir=/Users/pangxianhai/workspace/recruitment-manager
     cd /Users/pangxianhai/projects/clinical-recruitment/recruitment-manager
     rm -rf ./dist/*
@@ -43,21 +26,27 @@ function build_manager() {
     cd ..
     tar_file=`find . -name '*.tar.gz'`
     if [ -d $build_dir ]; then
-    rm -rf $build_dir
+       rm -rf $build_dir
     fi
     mkdir -p $build_dir
     tar zxvf $tar_file -C $build_dir
 }
 
-
-
-function work_build_web() {
-    source_dir=/Users/pangxianhai/projects/clinical-recruitment/recruitment-web
-    build $source_dir recruitment-web /Users/pangxianhai/workspace
-
+function build_server_dev() {
+    cd /Users/pangxianhai/projects/clinical-recruitment/recruitment-server
+    build_dir=/Users/pangxianhai/workspace/recruitment-server
+    gradle clean
+    gradle build -x test
+    supervisorctl stop recruitment-server
+    if [ -d $build_dir ]; then
+       rm -rf $build_dir
+    fi
+    mkdir -p $build_dir/logs
+    cp recruitment-web/build/libs/recruitment-server.jar $build_dir/
+    supervisorctl start recruitment-server
 }
 
-function build_product() {
+function build_server_product() {
     cd /Users/pangxianhai/projects/clinical-recruitment/recruitment-web
     mvn -Dmaven.test.skip=true -Pproduct package
     scp recruitment-boot/target/recruitment-boot-product.tar.gz root@59.110.230.31:/home/release
@@ -89,19 +78,18 @@ function build_manager_product() {
 }
 
 
-if [ "$1" == "web" ];then
-    work_build_web
-elif [ "$1" == "pro" ];then
-    build_product
-elif [ "$1" == "view" ];then
-    build_view
+if [ "$1" == "server" ];then
+    build_server_dev
+elif [ "$1" == "server-pro" ];then
+    build_server_product
 elif [ "$1" == "manager" ];then
-	build_manager
-elif [ "$1" == "pro-view" ];then
+    build_manager_dev
+elif [ "$1" == "manager-pro" ];then
+	build_manager_product
+elif [ "$1" == "view" ];then
+    build_view_dev
+elif [ "$1" == "view-pro" ];then
     build_view_product
-elif [ "$1" == "pro-manager" ];then
-    build_manager_product
 else
-    echo $0 "[web|pro|view|manager|pro-view|pro-manager]"
+    echo $0 "[server|server-pro|manager|manager-pro|view|view-pro]"
 fi
-
